feat: Add CfpRepo
