+++ internal/cfp/domain/repo.go
@@ -2,10 +2,11 @@ package domain
 
 import (
 	"context"
-	"dreamkast-weaver/internal/dkui/value"
+	"dreamkast-weaver/internal/cfp/value"
+	"net"
 )
 
 type CfpRepo interface {
 	ListCfpVotes(ctx context.Context, confName value.ConfName) (*CfpVotes, error)
-	InsertCfpVote(ctx context.Context, cfpVote *CfpVote) error
+	InsertCfpVote(ctx context.Context, confName value.ConfName, talkID value.TalkID, clientIp net.IP) error
 }
